Move E2E test codes into tests/test.rs
